move encryption methods to `include`
move files related to encryption and decryption to `include/encryption.php`
